Fix broken citation rendering
Previously, in-text citations consisting of only an `@` (not surrounded by [brackets])
didn't render at all -- they just completely disappeared after the cursor was no longer
on top of them.

Turns out this was because we recently introduced another capture group in the
citationRE -- so we have to skip that capture group when refering to it later on.