Disable pref if AppIndicator Support not detected
If running Linux and Gnome Desktop and AppIndicator Support [1] support
is not detected, disable the preference `Leave app running in the
notification area` and set to false. Style the disabled preference
(Linux only). Style is suitable for normal mode and dark mode.
Uses an environment variable as suggested by @nathanlesage [2]

[1] https://extensions.gnome.org/extension/615/appindicator-support/
[2] https://github.com/Zettlr/Zettlr/issues/175#issuecomment-814900960