perf(file-manager): Only show tree items if visible
Small but huge: This commit switches the rendering of tree
items that are not visible completely off (v-if instead of v-show).
This dramatically increases the performance of both filtering and
display of tree items with arbitrary sizes of the full file tree. The
only caveat is that the collapsed-status is not retained as of now.