Merge contributing guidelines from CONTRIBUTING.md into the README.md file
Guidelines on how to contribute to the development of Zettlr could previously
be found in two different places: the CONTRUBUTING.md file and the Contributing
section in the main README.md file. To avoid further fragmentation of information,
and to make it easier for new contributors to learn about the best way to help
Zettlr, I've merged the information from CONTRIBUTING.md into README.md. I made
sure to remove any duplicate information and only keep that information that
is the most up-to-date or the most complete. Additional changes were made to
improve the clarity of the README.md file's text.