build(i18n): Don't run i18n workflow via cron
This commit disables the cron for the i18n workflow.
Instead, we can now manually trigger it, and it will
automatically run whenever a new proper release is
being crafted. Since the translations will work even with
outdated PO-files (in English) and most community
translations only come after new releases, this is a good
balance between keeping the translations up to date
without causing this many merge conflicts that we've had
in the past.