fix: Table Node parser
What this commit effectively does is rewrite the table node
parser logic in such a way that it is no longer reliant on the
underlying Lezer parser's ability to properly produce
TableCell nodes. This makes the code a bit more clunky and
we hopefully won't fall into any performance traps, but in
theory this is now more precise, and even supports grid tables!