fix: Tweak citation selection to be less greedy
Fixes #5192

When clicking on a thing -- such as a citation -- that has a (soft)
line-wrap in the middle of the rendered text the current (unwanted)
behaviour is that the entirety of all lines that the clicked-on
thing spans are selected.

This fixes the behaviour so that *only* the clicked-on thing is
selected, even when the clicked-on thing spans multiple lines due to
a (soft) line-wrap.