feat: Cache ESLint to improve speed
Running ESLint on the command line is slow.

This change enables ESLint caching which speeds up subsequent runs of the tool
on the command line.

* [Caching](https://eslint.org/docs/latest/use/command-line-interface#caching)

Current timings with no caching.

```
real    1m32.620s
```

After change, with caching.

```
real    0m7.296s
```

I am using the ESLint extension for in-the-moment linting of files I'm working
on. Still, I lilke to run the `yarn lint` and `yarn test` scripts before
pushing just to ensure that I haven't b0rk3d anything.

Waiting near two minutes for the linting to complete is punishing. The caching
is not a great solution -- waiting near 10 seconds is still pretty slow -- but
it's a marked improvement on the performance without it.

Note that this doesn't have any impact on the GitHub Actions builds; they are
still taking near two minutes to lint the codebase.

* [Recent build](https://github.com/Zettlr/Zettlr/actions/runs/14171139998/job/39695001535)

Addressing the GitHub Actions build times is a separate issue I think; this
fix is all about improving (one small aspect of) the Developer Experience
(DX) of working on the Zettlr codebase.