feat: Add transform to change italics to quotes
This PR addresses half of #5677: the code for doing the reverse transformation — quotes to italics — will follow in a distinct PR.

This adds another transform to the editor's new _Transform_ submenu which will change italicized text to quoted text.

In the video I am:

* selecting some italicized text
* clicking the command from the aforementioned submenu

After running the command the italicized text becomes instead quoted text.

Tested on: Windows 11 Home.

> Note that I have not updated the CHANGELOG per the PR template guideline 'cos I don't know when this'll be merged, if ever 🤷