fix: toSentenceCase no longer skips the first word
Given a sentence where the first word was ALL_CAPS, the `toSentenceCase` function would transform the other, subsequent, words in the sentence to sentence case... but it'd essentially skip casing the first word.

This fix adds a test case to flush out the bug and then fixes the bug in the implementation code.