fix: Consistent layout of toolbar on Windows
The buttons to search for files and toggle the file manager are stacking vertically when the window is shrunk horizontally. This doesn't match the behaviour of those buttons on other platforms where no such stacking occurs.

This change updates the CSS so that the layout behaviour is consistent -- no stacking -- across platforms.